# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Interview Position Tracker API",
    "description": "\n    A comprehensive REST API for tracking job positions and interview progress.\n    \n    ## Features\n    \n    * **User Authentication**: Secure JWT-based authentication system\n    * **Position Management**: Create, read, update, and delete job positions\n    * **Interview Tracking**: Track multiple interview stages for each position\n    * **Statistics & Analytics**: Get insights into your job search progress\n    * **Data Filtering**: Filter positions by status, company, date range, and more\n    * **Secure Access**: User-specific data isolation and authorization\n    \n    ## Authentication\n    \n    Most endpoints require authentication. To authenticate:\n    \n    1. Register a new account using `/api/v1/auth/register`\n    2. Login using `/api/v1/auth/login` to get an access token\n    3. Include the token in the Authorization header: `Bearer <your_token>`\n    \n    ## Workflow\n    \n    1. **Register/Login** to get authenticated\n    2. **Create positions** for jobs you're applying to\n    3. **Add interviews** as you progress through hiring processes\n    4. **Update statuses** as your applications move forward\n    5. **View statistics** to analyze your job search performance\n    ",
    "contact": {
      "name": "Interview Position Tracker API",
      "url": "https://github.com/your-repo/interview-tracker"
    },
    "license": {
      "name": "MIT",
...
  },
  "paths": {
...
    "/api/v1/positions/": {
      "post": {
        "tags": [
          "Positions"
        ],
        "summary": "Create a new job position",
        "description": "Create a new job position to track in your application process.",
        "operationId": "create_position_api_v1_positions__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Position successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                },
                "example": {
                  "id": "123e4567-e89b-12d3-a456-426614174000",
                  "title": "Senior Software Engineer",
                  "company": "TechCorp",
                  "description": "Full-stack development with React and Python",
                  "location": "San Francisco, CA",
                  "salary_range": "$120k - $180k",
                  "status": "applied",
                  "application_date": "2024-01-15",
                  "created_at": "2024-01-15T10:30:00Z",
                  "updated_at": "2024-01-15T10:30:00Z",
                  "interviews": []
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "tags": [
          "Positions"
        ],
        "summary": "List Positions",
        "description": "List all positions for the authenticated user.\n\nSupports filtering by status, company, date range, and search terms.\nResults are paginated and can be sorted by various fields.",
        "operationId": "list_positions_api_v1_positions__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PositionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by position status",
              "title": "Status"
            },
            "description": "Filter by position status"
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by company name (partial match)",
              "title": "Company"
            },
            "description": "Filter by company name (partial match)"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter positions from this application date",
              "title": "Date From"
            },
            "description": "Filter positions from this application date"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter positions up to this application date",
              "title": "Date To"
            },
            "description": "Filter positions up to this application date"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in title, company, or description",
              "title": "Search"
            },
            "description": "Search in title, company, or description"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Field to sort by",
              "default": "application_date",
              "title": "Sort By"
            },
            "description": "Field to sort by"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort order: asc or desc",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order: asc or desc"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/positions/{position_id}": {
      "get": {
        "tags": [
          "Positions"
        ],
        "summary": "Get Position",
        "description": "Get a specific position by ID.\n\nReturns the position details if it exists and belongs to the authenticated user.",
        "operationId": "get_position_api_v1_positions__position_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Positions"
        ],
        "summary": "Update Position",
        "description": "Update a specific position.\n\nUpdates the position with the provided data if it exists and belongs to the authenticated user.\nOnly provided fields will be updated.",
        "operationId": "update_position_api_v1_positions__position_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Position Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Positions"
        ],
        "summary": "Delete Position",
        "description": "Delete a specific position.\n\nDeletes the position and all associated interview records if it exists and belongs to the authenticated user.",
        "operationId": "delete_position_api_v1_positions__position_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/positions/{position_id}/status": {
      "put": {
        "tags": [
          "Positions"
        ],
        "summary": "Update Position Status",
        "description": "Update the status of a specific position.\n\nUpdates only the status field of the position if it exists and belongs to the authenticated user.\nThis is a convenient endpoint for quick status updates without needing to send all position data.",
        "operationId": "update_position_status_api_v1_positions__position_id__status_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Position Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewPlace": {
        "type": "string",
        "enum": [
          "phone",
          "video",
          "onsite"
        ],
        "title": "InterviewPlace",
        "description": "Format/location where the interview takes place."
      },
      "InterviewResponse": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/InterviewType",
            "description": "Type of interview"
          },
          "place": {
            "$ref": "#/components/schemas/InterviewPlace",
            "description": "Format/location of interview"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Date",
            "description": "Scheduled date and time of interview"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes",
            "description": "Interview duration in minutes"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Additional notes about the interview"
          },
          "outcome": {
            "$ref": "#/components/schemas/InterviewOutcome",
            "description": "Interview outcome",
            "default": "pending"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Interview unique identifier"
          },
          "position_id": {
            "type": "string",
            "format": "uuid",
            "title": "Position Id",
            "description": "Associated position identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Interview creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Interview last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "type",
          "place",
          "scheduled_date",
          "id",
          "position_id",
          "created_at",
          "updated_at"
        ],
        "title": "InterviewResponse",
        "description": "Schema for interview response data."
      },
...
      "InterviewType": {
        "type": "string",
        "enum": [
          "technical",
          "behavioral",
          "hr",
          "final"
        ],
        "title": "InterviewType",
        "description": "Type of interview based on its purpose."
      },
...
      "PositionCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "title": "Title",
            "description": "Job position title"
          },
          "company": {
            "type": "string",
            "maxLength": 255,
            "title": "Company",
            "description": "Company name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Job description"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "Job location"
          },
          "salary_range": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Range",
            "description": "Salary range"
          },
          "status": {
            "$ref": "#/components/schemas/PositionStatus",
            "description": "Application status",
            "default": "applied"
          },
          "application_date": {
            "type": "string",
            "format": "date",
            "title": "Application Date",
            "description": "Date when application was submitted"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company",
          "application_date"
        ],
        "title": "PositionCreate",
        "description": "Schema for creating a new position."
      },
      "PositionListResponse": {
        "properties": {
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionResponse"
            },
            "type": "array",
            "title": "Positions",
            "description": "List of positions"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of positions"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page number"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page",
            "description": "Number of items per page"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "description": "Whether there are more pages"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev",
            "description": "Whether there are previous pages"
          }
        },
        "type": "object",
        "required": [
          "positions",
          "total",
          "page",
          "per_page",
          "has_next",
          "has_prev"
        ],
        "title": "PositionListResponse",
        "description": "Schema for paginated position list response."
      },
      "PositionResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "title": "Title",
            "description": "Job position title"
          },
          "company": {
            "type": "string",
            "maxLength": 255,
            "title": "Company",
            "description": "Company name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Job description"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "Job location"
          },
          "salary_range": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Range",
            "description": "Salary range"
          },
          "status": {
            "$ref": "#/components/schemas/PositionStatus",
            "description": "Application status",
            "default": "applied"
          },
          "application_date": {
            "type": "string",
            "format": "date",
            "title": "Application Date",
            "description": "Date when application was submitted"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Position unique identifier"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "Owner user identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Position creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Position last update timestamp"
          },
          "interviews": {
            "items": {
              "$ref": "#/components/schemas/InterviewResponse"
            },
            "type": "array",
            "title": "Interviews",
            "description": "Associated interviews",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "company",
          "application_date",
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "PositionResponse",
        "description": "Schema for position response data."
      },
      "PositionStatus": {
        "type": "string",
        "enum": [
          "applied",
          "screening",
          "interviewing",
          "offer",
          "rejected",
          "withdrawn"
        ],
        "title": "PositionStatus",
        "description": "Status of a job position application."
      },
      "PositionStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/PositionStatus",
            "description": "New application status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "PositionStatusUpdate",
        "description": "Schema for updating position status only."
      },
      "PositionUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Job position title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Company",
            "description": "Company name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Job description"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "Job location"
          },
          "salary_range": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Range",
            "description": "Salary range"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "Application status"
          },
          "application_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Date",
            "description": "Date when application was submitted"
          }
        },
        "type": "object",
        "title": "PositionUpdate",
        "description": "Schema for updating an existing position."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}